Merge branch 'fix/FCduplicateImport' into 'master'
Avoid reading fld twice in FieldConvert

This MR introduces a small change to avoid loading the fld file twice in FieldConvert. Currently, InputXml imports just the FieldDef to set the expansion, and later InputFld reads the data. However, this is very inefficient, since the whole fld file has to be loaded even if we just want the FieldDef. This MR basically makes InputXml load both the fielddef and the data, and later InputFld just extracts the data.

See merge request !649