fs check logic fix for hdf5
in case of perRank=false, we dont need to check if the specPath is
already gone for all ranks since only rank0 will create the new file
anyway.

This should avoid polling the filesystem from every rank when hdf5 is used.

Just for correctness: the specPath is a file, not a dir, so change the test
accordingly