Updated Timer class to prevent wrong measurements
Updated the Timer class such that it imposes the following requirements:
    - Timer::Start() can only be called if the timer is not running
    - Timer::Stop() can only be cauued if the timer is running
    - Timer::Elapsed() can only be called if the timer is not running

The status of the timer is checked with a private member variable "m_isactive"

The m_start and n_end member variables have also been initialized to 0. This ensures
that Timer::Elapsed() will return 0 seconds, in case it is called before the timer
was started+stopped. That is, it is possible to create a timer, and then call
Timer::Elapsed(), and get 0 seconds back. If the m_start and m_end variables are
left un-initialized, there is a risk that we get arbitrary values from Timer::Elapsed().