Fixed bugs when timing the code
Found two places where the Timer class was used incorrectly to time
the execution of different routines.

In ExpList.cpp, the timer was started inside an if statement, which
means that it will not always start.

In DiffusionIP.cpp, the timer was not restarted after it was stopped.

The fixes make sure that reasonable values are printed after the code
completes. Before, values in the order of 1E8 were printed.