Improved formatting for Timer class
Updated the Timer class to get nicer formatting of results.

Removed static variable m_maxStringWidth and instead calculated
the width of each column inside the print routine. There is no
point in updating a static member variable on the fly, when it
is only needed once at the end to print the results.