Updated Copy Assignment in PushBackFunctor
Since the copy assignment operator (operator=) was already designed
to not allow assignment, it has been explicitly deleted. This way,
there is no risk that copy assignment happens. With this change,
the code compiles on ARCHER2 with Cray build environment.

There should be no risk that this commit changes the behaviour of
the code, since we only removed a function. If this function would
have been used before, the code would not have compiled.

The compiler still generates an implicit move assignment operator,
copy constructor, and move constructor. These can not be deleted
(if this is done, the code does not compile, indicating that they
are used somewhere). Since these operators are not explicitly defined,
we can't say for sure that the constructors and move assignment
operators do what we excpect them to do.