Updated timers in CompressibleFlowSolver
Removed a couple of timers.

Changed iolevel to 10 for some timers that we are not sure how
important they are for the moment. All timers with iolevel=10
are not necessarily unimportant, but their relative importance
is not clear for the moment, so we have decided to make them
iolevel=10 to avoid them being printed.