Updated timers in CompressibleFlowSolver
Changed timers with iolevel=3 to iolevel=10. This is not because these
timers are not important, but because we don't know the relative importance
of these timers at the moment. Therefore, we decided to put them at a high
iolevel to avoid them being printed by a user. Once we know move about which
functions that are important, we can set these iolevels to a more reasonable
value.