Working on selective HDF5 reading of fld file
Fixed more compiler errors due to type missmatch.

Code compiles now if we use "boost::ignore_unused()" on a few variables
inside ImportFieldData(). However, we should fix the parameter list of
this function instead, since it does not need all the variable. Alternatively,
we might want to look at passing some data to this function. For example,
we could always pass the dataIndxToRead vector and not differentiate
between selective reading and not.