Working on selective HDF5 reading of fld file
Fixed so that the selective data indices that point into the plain
data structure that we store in the HDF5 file reads all the fields,
i.e., all the variables stored in each FieldDefinition. Before, we
only read the first field, which caused an error.

Also found a few places where we might run into problems, tried to
add comments to these places. In particular, what happens when we
don't have a uniform polynomial order, and run CheckFieldDef on the
new FieldDefinition (with updated m_elementIDs)? All the remaining
metadata in the FieldDefinition is untouched, such as number of modes
(coefficients) per element.