Working on selective HDF5 reading of fld file
Fixed error when computing the indices in the data array that should
be read. I had forgotten to increase the offset each time a new variable
was read. After the fix, the offset into the dataIdxToRead array is shifted
by "offset", where offset = #elements * #modes per element in the
decomposition.