Working on selective HDF5 reading of fld file
Moved the logic for computing the number of data points needed to
store the elemental expansion for an arbitrary element outside
the routine FieldIO::CheckFieldDefinition(). This is done to avoid
duplicate code, which occurs because we do the same calculation
for both uniform and variable order elements. I.e., when all elements
in an expansion have the same polynomial order, and when the elements
can have a different polynomial order.