Working on selective HDF5 reading of fld file
Added separate function for computing the number of coefficients
per element. This functionality was previously added to the
FieldIO::CheckFieldDefinition() function, but it was agreed that
it is better to move this functionality to a separate function.
In the next commit, the FieldIO::CheckFieldDefinition() function
will be restored to its previous state.