Working on selective HDF5 reading of fld file
Added updating of counter that was missing for Quad and Tri elements
when one homogenous direction is used. This counter was missing because
the function GetNumberOfDataPoints() contains the code that was previously
located in CheckFieldDefinition(). In the latter function, there was an
if statement that checked whether a uniform polynomial order was used
or not. The code that is now in GetNumberOfDataPoints() was then duplicated
in both branches of the if-statement, and I therefore copied one of the
branches into the new function GetNumberOfDataPoints. Hoever, there was
a small discrepancy between the two branches, namely that the counter
was not updated for the uniform order case. This was not added there
because it was not needed, since only the first element is read for
uniform order cases. However, it can be added without changing anything.
I have now added it to the function GetNumberOfDataPoints(), which
replaces the code inside the two if-branches. I have verified that this
was the only difference between the two if-branches.