Working on selective HDF5 reading of fld file
Restored FieldIO::CheckFieldDefinition() to its original functionality, except
that the function now calls an external function called GetNumberOfDataPoints()
to compute the number of values needed to store the polynomial expansion inside
each element. Before, this code was explicitly included in CheckFieldDefinition().

To restore the function, I removed the additional std::vector that was passed to
the function. In a previous commit, this vector was filled by CheckFieldDefinition(),
which meant that CheckFieldDefinition() served two purposes: fill the vector with
data, and check that the FieldDefinition that is passed to the function is correct.
It was agreed that the std::vector should be filled up by a separate function instead.
This function is called GetNumberOfCoeffsPerElmt(), and it returns a std::vector. By
relying on move semantics, this can be done efficiently without any deep copy.